Ship refactor
Split ship into seperate components. Upgrades now in a static class.
