+++ Assets/Scripts/Misc/CameraMoveWithShip.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: b9162431f2bb9144194a8843e8130776
+timeCreated: 1526818220
+licenseType: Pro
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
