+++ Assets/Scripts/Player/Ship.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: dfddc87cd7be4814ca9316e5622cad55
+folderAsset: yes
+timeCreated: 1526824503
+licenseType: Pro
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
