+++ Assets/Scripts/Player/Ship/ShipBuffs.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 4bea307ff014a964b98f12d098880914
+timeCreated: 1526818085
+licenseType: Pro
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
