+++ Assets/Scripts/Player/Ship/ShipShoot.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 390c58bc80eaf4a469b1f7c0b4fb5122
+timeCreated: 1526817418
+licenseType: Pro
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
