+++ Assets/Scripts/Player/Ship/StartingShip.cs.meta
