+++ Assets/Scripts/Player/Upgrades.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 0e63afad15f166143a130b94b7072d41
+timeCreated: 1526817678
+licenseType: Pro
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
